Fix UB: avoid calls to memcpy(0, 0, 0)
It is undefined behavior to call memcpy with null pointers,
even with 0 size/count. Thus we introduce a function
safer_memcpy which wraps std::copy in such a way that
it is safe also to call with null pointers when size/count == 0.
Since we only use std::copy on a char pointers, the compiler
should be able to produce the same code, but without going
down the route of undefined behavior.