Rewrite binaryOpSparseNewTest
The logic of the test was very complicated, and in fact
fsanitize=address as well as valgrind reported buffer overflows
due to invalid memory accesses. Instead of repairing the logic,
the test has been rewritten such that it uses more direct knowledge
of the expected sparsity pattern of the result, and it is now more
straightforward to understand what is actually being tested.