Implement host-versions of access classes
This allows us to provide an iterator/STL-compatible
interface as Host{Read/Write/WriteOnly}Access is guaranteed
to always have its memory reside on the host, and hence
STL algorithms are applicable.

In the same commit, we changed the writeOnlyAccess() functions
to require a newSize being provided (WriteOnlyAccess will probably
be changed in the future so that the constructors that do not
take a new size get removed).